Move persisted log into data plugin / query
