+++ src/legacy/core_plugins/data/public/plugin.ts
@@ -107,6 +107,7 @@ export class DataPlugin
 
     const timefilterService = this.timefilter.setup({
       uiSettings,
+      store: __LEGACY.storage,
     });
     this.setupApi = {
       indexPatterns: indexPatternsService,
