+++ src/legacy/core_plugins/data/public/query/query_bar/components/query_bar_input.tsx
@@ -25,7 +25,6 @@ import { EuiFieldText, EuiOutsideClickDetector, PopoverAnchorPosition } from '@e
 
 import { InjectedIntl, injectI18n } from '@kbn/i18n/react';
 import { debounce, compact, isEqual } from 'lodash';
-import { PersistedLog } from 'ui/persisted_log';
 
 import {
   AutocompleteSuggestion,
@@ -40,7 +39,7 @@ import { Query } from '../index';
 import { fromUser, matchPairs, toUser } from '../lib';
 import { QueryLanguageSwitcher } from './language_switcher';
 import { SuggestionsComponent } from './typeahead/suggestions_component';
-import { getQueryLog } from '../lib/get_query_log';
+import { PersistedLog, getQueryLog } from '../../persisted_log';
 import { fetchIndexPatterns } from '../lib/fetch_index_patterns';
 import { IDataPluginServices } from '../../../types';
 
@@ -392,6 +391,7 @@ export class QueryBarInputUI extends Component<Props, State> {
   };
 
   public componentDidMount() {
+    const { uiSettings, store, appName } = this.services;
     const parsedQuery = fromUser(toUser(this.props.query.query));
     if (!isEqual(this.props.query.query, parsedQuery)) {
       this.onChange({ ...this.props.query, query: parsedQuery });
@@ -399,12 +399,13 @@ export class QueryBarInputUI extends Component<Props, State> {
 
     this.persistedLog = this.props.persistedLog
       ? this.props.persistedLog
-      : getQueryLog(this.services.uiSettings, this.services.appName, this.props.query.language);
+      : getQueryLog(uiSettings, store, appName, this.props.query.language);
 
     this.fetchIndexPatterns().then(this.updateSuggestions);
   }
 
   public componentDidUpdate(prevProps: Props) {
+    const { uiSettings, store, appName } = this.services;
     const parsedQuery = fromUser(toUser(this.props.query.query));
     if (!isEqual(this.props.query.query, parsedQuery)) {
       this.onChange({ ...this.props.query, query: parsedQuery });
@@ -412,7 +413,7 @@ export class QueryBarInputUI extends Component<Props, State> {
 
     this.persistedLog = this.props.persistedLog
       ? this.props.persistedLog
-      : getQueryLog(this.services.uiSettings, this.services.appName, this.props.query.language);
+      : getQueryLog(uiSettings, store, appName, this.props.query.language);
 
     if (!isEqual(prevProps.indexPatterns, this.props.indexPatterns)) {
       this.fetchIndexPatterns().then(this.updateSuggestions);
